cli: rename structured command package
